+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/settings/AppSettingsConfigurable.kt
@@ -6,8 +6,8 @@ import com.github.blarc.ai.commits.intellij.plugin.OpenAIService
 import com.intellij.icons.AllIcons
 import com.intellij.openapi.options.BoundConfigurable
 import com.intellij.openapi.progress.runBackgroundableTask
-import com.intellij.openapi.ui.naturalSorted
 import com.intellij.ui.components.JBLabel
+import com.intellij.ui.components.JBTextArea
 import com.intellij.ui.dsl.builder.*
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.Dispatchers
@@ -20,6 +20,12 @@ class AppSettingsConfigurable : BoundConfigurable(message("settings.general.grou
 
     private val tokenPasswordField = JPasswordField()
     private val verifyLabel = JBLabel()
+    private val promptTextArea = JBTextArea()
+    init {
+        promptTextArea.wrapStyleWord = true
+        promptTextArea.lineWrap = true
+        promptTextArea.isEditable = false
+    }
     override fun createPanel() = panel {
 
         row {
@@ -47,6 +53,21 @@ class AppSettingsConfigurable : BoundConfigurable(message("settings.general.grou
                 .label(message("settings.locale"))
                 .bindItem(AppSettings.instance::locale.toNullableProperty())
         }
+        row {
+            comboBox(AppSettings.instance.prompts.keys.toList(), AppSettingsListCellRenderer())
+                .label(message("settings.prompt"))
+                .bindItem(AppSettings.instance::currentPrompt.toNullableProperty())
+                .onChanged { promptTextArea.text = AppSettings.instance.prompts[it.item] }
+        }
+        row {
+            cell(promptTextArea)
+                .bindText(
+                    { AppSettings.instance.getPrompt("") },
+                    {  }
+                )
+                .align(Align.FILL)
+                .resizableColumn()
+        }.resizableRow()
     }
 
     @OptIn(DelicateCoroutinesApi::class)
@@ -61,7 +82,6 @@ class AppSettingsConfigurable : BoundConfigurable(message("settings.general.grou
 
                 GlobalScope.launch(Dispatchers.IO) {
                     try {
-                        println(String(tokenPasswordField.password))
                         OpenAIService.instance.verifyToken(String(tokenPasswordField.password))
                         verifyLabel.text = message("settings.verify.valid")
                         verifyLabel.icon = AllIcons.General.InspectionsOK